Apple Silicon (macOS/MPS) port: streamlined runner, docs, packaging
Run SAM 3D Objects on Apple Silicon end to end (no CUDA):

- run.sh: single streamlined flow (image -> splat -> textured GLB) in two
  separate processes so only one memory-heavy stage is resident at a time;
  first-launch model-weights check that prints HF download steps; keep
  `glb <dir>` to re-bake from an existing splat.
- cli.py: two-prompt flow (image path + output name -> outputs/<name>/),
  always rembg-extract, always full pipeline, random seed capped 0-41,
  NaN guard that refuses to write a dead splat.
- ply2glb.py: standalone splat -> textured GLB using only the mesh decoder.
- gsplat_silicon / mesh_raster_silicon: pure-PyTorch EWA + z-buffered
  rasterizers replacing CUDA gsplat / nvdiffrast; float64 math on CPU;
  fp32 mesh decoder.
- README rewritten for the macOS port (weights download, usage, internals);
  upstream README kept as README.upstream.md. LICENSE unchanged (Meta SAM).
- .gitignore: exclude weights (checkpoints/hf), outputs, venvs, caches.

+++ sam3d_objects/model/backbone/tdfy_dit/modules/sparse/conv/conv_native.py
@@ -75,7 +75,7 @@ def _subm_forward(
                     src_t = torch.tensor(src_idx, device=feats.device)
                     nb_t = torch.tensor(nb_idx, device=feats.device)
                     gathered = feats[nb_t]         # [M, Cin]
-                    contrib = gathered @ W          # [M, Cout]
+                    contrib = (gathered @ W).to(out.dtype)  # [M, Cout]
                     out.index_add_(0, src_t, contrib)
 
     if bias is not None:
@@ -169,7 +169,7 @@ def _sparse_forward(
                 if src_idx_list:
                     src_t = torch.tensor(src_idx_list, device=feats.device)
                     dst_t = torch.tensor(dst_idx_list, device=feats.device)
-                    contrib = feats[src_t] @ W   # [M, Cout]
+                    contrib = (feats[src_t] @ W).to(out_feats.dtype)   # [M, Cout]
                     out_feats.index_add_(0, dst_t, contrib)
 
     if bias is not None:
@@ -389,7 +389,7 @@ class SparseInverseConv3d(nn.Module):
                         src_t = torch.tensor(src_list, device=feats.device)
                         dst_t = torch.tensor(dst_list, device=feats.device)
                         # feats[src] @ W.T  => [M, Cout]
-                        contrib = feats[src_t] @ W.T
+                        contrib = (feats[src_t] @ W.T).to(out_feats.dtype)
                         out_feats.index_add_(0, dst_t, contrib)
 
         if self.bias is not None:
